requirements.txt
# =======================
# EXISTING DEPENDENCIES
# =======================
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
pytesseract==0.3.10
Pillow==9.3.0
Werkzeug==2.2.2
SQLAlchemy==1.4.42
python-dateutil==2.8.2
scikit-learn==1.2.0
pandas==1.5.2
Faker==15.3.4
qrcode==7.3.1
pypdf==3.17.0

# =======================
# NEW AI/ML DEPENDENCIES
# =======================
# AI APIs
openai==1.3.9
anthropic==0.7.4
google-generativeai==0.3.2

# Async support
aiohttp==3.8.6

# Flask extensions
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
WTForms==3.0.1
email-validator==2.1.0
phonenumbers==8.13.22

# PDF Processing (complementary to existing pypdf)
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Data Processing
numpy==1.25.2

# Security
bcrypt==4.0.1
cryptography==41.0.7

# Database (optional - for PostgreSQL)
psycopg2-binary==2.9.9